add new boundary

+++ Field.h
@@ -19,7 +19,7 @@ public:
     T sum() const{
         T s = 0;
         for(int d = 0; d < dim_; ++d){
-            for(int i = 0; i < data_[d].size(); ++i)
+            for(size_t i = 0; i < data_[d].size(); ++i)
                 s += data_[d][i];
         }
         return s;
